put_db_data/dummy2doc/requirements.txt
mysql-connector-python
python-dotenv
sqlalchemy
tqdm